[ fix ] API 호출 분리
