+++ src/@components/@common/MenuBar/style.ts
@@ -25,57 +25,6 @@ export const St = {
     top: 10rem;
   `,
 
-  ProfileContainer: styled.article`
-    border-bottom: solid 0.2rem #f1f1f1;
-
-    padding-bottom: 3.6rem;
-  `,
-
-  ProfileImage: styled.img`
-    width: 5.8rem;
-    height: 5.8rem;
-    border-radius: 50%;
-  `,
-
-  WelcomeText: styled.h1`
-    ${({ theme }) => theme.fonts.body10};
-    color: ${({ theme }) => theme.colors.bg};
-
-    margin-top: 1.6rem;
-  `,
-
-  DescriptText: styled.h2`
-    ${({ theme }) => theme.fonts.body12};
-    color: ${({ theme }) => theme.colors.gray600};
-
-    margin-top: 0.4rem;
-  `,
-
-  BtnContainer: styled.div`
-    display: flex;
-    align-items: center;
-
-    color: ${({ theme }) => theme.colors.gray600};
-
-    margin-top: 3.6rem;
-  `,
-
-  MyProfileBtn: styled(Link)`
-    ${({ theme }) => theme.fonts.btn2};
-    color: ${({ theme }) => theme.colors.gray600};
-
-    border-right: solid 0.15rem ${({ theme }) => theme.colors.gray600};
-
-    padding-right: 1rem;
-  `,
-
-  LogoutBtn: styled.button`
-    ${({ theme }) => theme.fonts.btn2};
-    color: ${({ theme }) => theme.colors.gray600};
-
-    padding-left: 1rem;
-  `,
-
   RecomendContainer: styled.article`
     padding-top: 3.6rem;
   `,
